pyproject.toml
[project]
name = "kairos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "colorama==0.4.6",
    "flask==3.1.0",
    "flask-login==0.6.3",
    "flask-restful>=0.3.10",
    "flask-wtf==1.2.2",
    "google-api-python-client==2.156.0",
    "lxml>=5.4.0",
    "mistralai==1.2.5",
    "openai>=1.78.0",
    "pydantic>=2.10.6",
    "python-dotenv==1.0.1",
    "sqlalchemy==2.0.38",
]

[dependency-groups]
testing = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "pytest>=8.3.5",
]
